+++ packages/editor/editor-core/example-helpers/ToolsDrawer.tsx
@@ -194,104 +194,109 @@ export default class ToolsDrawer extends React.Component<Props & any, State> {
     return (
       <AnalyticsListener channel="atlaskit" onEvent={e => console.log(e)}>
         <AnalyticsListener channel="media" onEvent={e => console.log(e)}>
-          <Content>
-            <div style={{ padding: '5px 0' }}>
-              ️️️⚠️ Atlassians, for Media integration to work in non-mocked
-              state, make sure you're logged into{' '}
-              <a href="https://id.stg.internal.atlassian.com" target="_blank">
-                staging Identity server.
-              </a>
-            </div>
-            {reloadEditor
-              ? ''
-              : this.props.renderEditor({
-                  disabled: !editorEnabled,
-                  imageUploadProvider:
-                    providers.imageUploadProvider[imageUploadProvider],
-                  mediaProvider: providers.mediaProvider[mediaProvider],
-                  mentionProvider: providers.mentionProvider[mentionProvider],
-                  emojiProvider: providers.emojiProvider[emojiProvider],
-                  taskDecisionProvider:
-                    providers.taskDecisionProvider[taskDecisionProvider],
-                  contextIdentifierProvider:
-                    providers.contextIdentifierProvider[
-                      contextIdentifierProvider
-                    ],
-                  activityProvider:
-                    providers.activityProvider[activityProvider],
-                  onChange: this.onChange,
-                })}
-            <div className="toolsDrawer">
-              {Object.keys(providers).map(providerKey => (
-                <div key={providerKey}>
+          <AnalyticsListener
+            channel="fabric-elements"
+            onEvent={e => console.log(e)}
+          >
+            <Content>
+              <div style={{ padding: '5px 0' }}>
+                ️️️⚠️ Atlassians, for Media integration to work in non-mocked
+                state, make sure you're logged into{' '}
+                <a href="https://id.stg.internal.atlassian.com" target="_blank">
+                  staging Identity server.
+                </a>
+              </div>
+              {reloadEditor
+                ? ''
+                : this.props.renderEditor({
+                    disabled: !editorEnabled,
+                    imageUploadProvider:
+                      providers.imageUploadProvider[imageUploadProvider],
+                    mediaProvider: providers.mediaProvider[mediaProvider],
+                    mentionProvider: providers.mentionProvider[mentionProvider],
+                    emojiProvider: providers.emojiProvider[emojiProvider],
+                    taskDecisionProvider:
+                      providers.taskDecisionProvider[taskDecisionProvider],
+                    contextIdentifierProvider:
+                      providers.contextIdentifierProvider[
+                        contextIdentifierProvider
+                      ],
+                    activityProvider:
+                      providers.activityProvider[activityProvider],
+                    onChange: this.onChange,
+                  })}
+              <div className="toolsDrawer">
+                {Object.keys(providers).map(providerKey => (
+                  <div key={providerKey}>
+                    <ButtonGroup>
+                      <label>{providerKey}: </label>
+                      {Object.keys(providers[providerKey]).map(
+                        providerStateName => (
+                          <Button
+                            key={`${providerKey}-${providerStateName}`}
+                            onClick={this.switchProvider.bind(
+                              this,
+                              providerKey,
+                              providerStateName,
+                            )}
+                            className={`${providerKey}-${providerStateName
+                              .replace(/[()]/g, '')
+                              .replace(/ /g, '-')}`}
+                            appearance={
+                              providerStateName === this.state[providerKey]
+                                ? 'primary'
+                                : 'default'
+                            }
+                            theme="dark"
+                            spacing="compact"
+                          >
+                            {providerStateName}
+                          </Button>
+                        ),
+                      )}
+                    </ButtonGroup>
+                  </div>
+                ))}
+                <div>
                   <ButtonGroup>
-                    <label>{providerKey}: </label>
-                    {Object.keys(providers[providerKey]).map(
-                      providerStateName => (
-                        <Button
-                          key={`${providerKey}-${providerStateName}`}
-                          onClick={this.switchProvider.bind(
-                            this,
-                            providerKey,
-                            providerStateName,
-                          )}
-                          className={`${providerKey}-${providerStateName
-                            .replace(/[()]/g, '')
-                            .replace(/ /g, '-')}`}
-                          appearance={
-                            providerStateName === this.state[providerKey]
-                              ? 'primary'
-                              : 'default'
-                          }
-                          theme="dark"
-                          spacing="compact"
-                        >
-                          {providerStateName}
-                        </Button>
-                      ),
-                    )}
-                  </ButtonGroup>
-                </div>
-              ))}
-              <div>
-                <ButtonGroup>
-                  <Button
-                    onClick={this.toggleDisabled}
-                    theme="dark"
-                    spacing="compact"
-                  >
-                    {this.state.editorEnabled
-                      ? 'Disable editor'
-                      : 'Enable editor'}
-                  </Button>
-                  <Button
-                    onClick={this.reloadEditor}
-                    theme="dark"
-                    spacing="compact"
-                    className="reloadEditorButton"
-                  >
-                    Reload Editor
-                  </Button>
-                  <Tooltip content="Hot reload is not supported. Enable or disable before opening media-picker">
                     <Button
-                      onClick={this.toggleMediaMock}
-                      appearance={mediaMockEnabled ? 'primary' : 'default'}
+                      onClick={this.toggleDisabled}
                       theme="dark"
                       spacing="compact"
-                      className="mediaPickerMock"
                     >
-                      {mediaMockEnabled ? 'Disable' : 'Enable'} Media-Picker
-                      Mock
+                      {this.state.editorEnabled
+                        ? 'Disable editor'
+                        : 'Enable editor'}
                     </Button>
-                  </Tooltip>
-                </ButtonGroup>
+                    <Button
+                      onClick={this.reloadEditor}
+                      theme="dark"
+                      spacing="compact"
+                      className="reloadEditorButton"
+                    >
+                      Reload Editor
+                    </Button>
+                    <Tooltip content="Hot reload is not supported. Enable or disable before opening media-picker">
+                      <Button
+                        onClick={this.toggleMediaMock}
+                        appearance={mediaMockEnabled ? 'primary' : 'default'}
+                        theme="dark"
+                        spacing="compact"
+                        className="mediaPickerMock"
+                      >
+                        {mediaMockEnabled ? 'Disable' : 'Enable'} Media-Picker
+                        Mock
+                      </Button>
+                    </Tooltip>
+                  </ButtonGroup>
+                </div>
               </div>
-            </div>
-            <div className="json-output">
-              <legend>JSON output:</legend>
-              <pre>{jsonDocument}</pre>
-            </div>
-          </Content>
+              <div className="json-output">
+                <legend>JSON output:</legend>
+                <pre>{jsonDocument}</pre>
+              </div>
+            </Content>
+          </AnalyticsListener>
         </AnalyticsListener>
       </AnalyticsListener>
     );