+++ packages/editor/editor-core/examples/5-full-page.tsx
@@ -133,7 +133,6 @@ export class ExampleEditor extends React.Component<EditorProps, State> {
             appearance="full-page"
             analyticsHandler={analyticsHandler}
             quickInsert={{ provider: Promise.resolve(quickInsertProvider) }}
-            delegateAnalyticsEvent={(...args) => console.log(args)}
             allowTasksAndDecisions={true}
             allowCodeBlocks={{ enableKeybindingsForIDE: true }}
             allowLists={true}