+++ packages/elements/util-data-test/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @atlaskit/util-data-test
 
+## 10.0.12
+- [patch] Updated dependencies [8b2c4d3](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/8b2c4d3)
+- [patch] Updated dependencies [3302d51](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/3302d51)
+  - @atlaskit/emoji@43.0.0
+  - @atlaskit/task-decision@10.0.1
+
+## 10.0.11
+- [patch] Updated dependencies [23c7eca](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/23c7eca)
+  - @atlaskit/task-decision@10.0.0
+
 ## 10.0.10
 - [patch] Updated dependencies [927ae63](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/927ae63)
   - @atlaskit/emoji@42.0.0