+++ packages/search/quick-search/CHANGELOG.md
@@ -1,5 +1,8 @@
 # @atlaskit/quick-search
 
+## 4.2.10
+- [patch] Fixing selected style for Item and Fixing focus on Quick search when component is not remounted [9532a1b](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/9532a1b)
+
 ## 4.2.9
 - [patch] Updated dependencies [65c6514](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/65c6514)
   - @atlaskit/docs@5.0.8